+++ Projet BDD (temp)/TABLES/07-PARLE.sql
@@ -5,7 +5,7 @@ DROP TABLE Parle;
 /*Table de lien entre Langues et Personnages*/
 CREATE TABLE  IF NOT EXISTS Parle(
     code VARCHAR(5),
-    id_perso TINYINT UNSIGNED,
+    id_perso SMALLINT UNSIGNED,
     
     CONSTRAINT FK_code FOREIGN KEY(code) REFERENCES Langues(code) NOT NULL,
     CONSTRAINT FK_perso FOREIGN KEY(id_perso) REFERENCES Personnages(id_personnage) NOT NULL,
